+++ .bumpversion.cfg
@@ -8,7 +8,7 @@ parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = 
 	{major}.{minor}.{patch}
 
-[bumpversion:file:jsonschema_spec/__init__.py]
+[bumpversion:file:jsonschema_path/__init__.py]
 
 [bumpversion:file:pyproject.toml]
 search = version = "{current_version}"
