+++ jsonschema_path/__init__.py
@@ -1,0 +1,11 @@
+from jsonschema_path.accessors import SchemaAccessor
+from jsonschema_path.handlers import default_handlers
+from jsonschema_path.paths import SchemaPath
+
+__author__ = "Artur Maciag"
+__email__ = "maciag.artur@gmail.com"
+__version__ = "0.2.4"
+__url__ = "https://github.com/p1c2u/jsonschema-path"
+__license__ = "Apache-2.0"
+
+__all__ = ["SchemaAccessor", "SchemaPath", "default_handlers"]
