+++ jsonschema_path/handlers/requests.py
@@ -7,9 +7,9 @@ from typing import Union
 
 import requests
 
-from jsonschema_spec.handlers.file import BaseFilePathHandler
-from jsonschema_spec.handlers.file import FileHandler
-from jsonschema_spec.handlers.protocols import SupportsRead
+from jsonschema_path.handlers.file import BaseFilePathHandler
+from jsonschema_path.handlers.file import FileHandler
+from jsonschema_path.handlers.protocols import SupportsRead
 
 
 class UrlRequestsHandler(BaseFilePathHandler):
