+++ jsonschema_path/handlers/utils.py
