+++ jsonschema_path/utils.py
