+++ pyproject.toml
@@ -4,13 +4,13 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.coverage.run]
 branch = true
-source =["jsonschema_spec"]
+source =["jsonschema_path"]
 
 [tool.coverage.xml]
 output = "reports/coverage.xml"
 
 [tool.mypy]
-files = "jsonschema_spec"
+files = "jsonschema_path"
 strict = true
 
 [[tool.mypy.overrides]]
@@ -18,13 +18,13 @@ module = "jsonschema_specifications"
 ignore_missing_imports = true
 
 [tool.poetry]
-name = "jsonschema-spec"
+name = "jsonschema-path"
 version = "0.2.4"
 description = "JSONSchema Spec with object-oriented paths"
 authors = ["Artur Maciag <maciag.artur@gmail.com>"]
 license = "Apache-2.0"
 readme = "README.rst"
-repository = "https://github.com/p1c2u/jsonschema-spec"
+repository = "https://github.com/p1c2u/jsonschema-path"
 keywords = ["jsonschema", "swagger", "spec"]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -58,7 +58,7 @@ flynt = "1.0.1"
 mypy = "^1.5.1"
 types-PyYAML = "^6.0.12"
 types-requests = "^2.31.0"
-typing-extensions = "^4.8.0"  # required by responses. See https://github.com/p1c2u/jsonschema-spec/issues/44
+typing-extensions = "^4.8.0"  # required by responses. See https://github.com/p1c2u/jsonschema-path/issues/44
 responses = "^0.23.3"
 deptry = "^0.12.0"
 
@@ -68,7 +68,7 @@ addopts = """
 --verbose
 --showlocals
 --junitxml=reports/junit.xml
---cov=jsonschema_spec
+--cov=jsonschema_path
 --cov-report=term-missing
 --cov-report=xml
 """
